requirements.txt
alembic==1.4.3
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
flask-oidc==1.4.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
httplib2==0.18.1
idna==2.10
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.4
MarkupSafe==1.1.1
oauth2client==4.1.3
Pillow==8.1.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
PyJWT==2.0.0
python-dateutil==2.8.1
python-editor==1.0.4
python-slugify==4.0.1
reportlab==3.5.59
requests==2.25.1
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.22
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
WTForms==2.3.3
Flask-Script==2.0.6